+++ MapboxNavigation/RouteMapViewController.swift
@@ -9,11 +9,6 @@ import SDWebImage
 class ArrowFillPolyline: MGLPolylineFeature {}
 class ArrowStrokePolyline: ArrowFillPolyline {}
 
-class DebugAnnotation: MGLPointAnnotation {
-    var color: UIColor = .red
-}
-class DebugAnnotationView: MGLAnnotationView {}
-
 class RouteMapViewController: UIViewController, PulleyPrimaryContentControllerDelegate {
     @IBOutlet weak var mapView: NavigationMapView!
 
@@ -275,40 +270,6 @@ class RouteMapViewController: UIViewController, PulleyPrimaryContentControllerDe
             }
         }
     }
-    
-    func addDebugAnnotation(location: CLLocation) {
-        let annotation = DebugAnnotation()
-        annotation.coordinate = location.coordinate
-        annotation.title = "Reroute"
-        annotation.color = .red
-        let newLocation = location.advanced(by: RouteControllerDeadReckoningTimeInterval)
-        let radius = location.rerouteRadius
-
-        var subtitle = location.debugInformation
-
-        subtitle += "\nMeters in front of user: \(location.distance(from: newLocation))"
-        subtitle += "\nRadius: \(radius)"
-        annotation.subtitle = subtitle
-        
-        let annotationInFrontOfUser = DebugAnnotation()
-        annotationInFrontOfUser.coordinate = newLocation.coordinate
-        annotationInFrontOfUser.title = "Dead reckoning"
-        annotationInFrontOfUser.subtitle = newLocation.debugInformation
-        annotationInFrontOfUser.color = .blue
-        mapView?.addAnnotation(annotation)
-        mapView?.addAnnotation(annotationInFrontOfUser)
-        
-        guard let closestCoordinate = closestCoordinate(on: routeController.routeProgress.currentLegProgress.currentStep.coordinates!, to: newLocation.coordinate) else {
-            return
-        }
-        
-        let closestAnnotation = DebugAnnotation()
-        closestAnnotation.coordinate = closestCoordinate.coordinate
-        closestAnnotation.title = "Closest coordinate"
-        closestAnnotation.subtitle = "Distance: \(closestCoordinate.distance)"
-        closestAnnotation.color = .green
-        mapView?.addAnnotation(closestAnnotation)
-    }
 }
 
 // MARK: NavigationMapViewDelegate
@@ -500,16 +461,12 @@ extension RouteMapViewController: MGLMapViewDelegate {
 
     func mapView(_ mapView: MGLMapView, didSelect annotation: MGLAnnotation) {
         if let annotation = annotation as? DebugAnnotation {
-            let controller = UIAlertController(title: annotation.title, message: annotation.subtitle, preferredStyle: .alert)
-            controller.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action) in
-                controller.dismiss(animated: true, completion: nil)
-            }))
-            present(controller, animated: true, completion: nil)
+            showDebugAlert(with: annotation)
             return
         }
         
         if resetTrackingModeTimer != nil {
-            resetTrackingModeTimer.invalidate()
+            resetTrackingModeTimer?.invalidate()
             startResetTrackingModeTimer()
         }
     }
@@ -520,20 +477,7 @@ extension RouteMapViewController: MGLMapViewDelegate {
     
     func mapView(_ mapView: MGLMapView, viewFor annotation: MGLAnnotation) -> MGLAnnotationView? {
         guard let debugAnnotation = annotation as? DebugAnnotation else { return nil }
-        var annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: "debugView")
-        if annotationView == nil {
-            annotationView = DebugAnnotationView(reuseIdentifier: "debugView")
-            annotationView?.frame = CGRect(origin: .zero, size: CGSize(width: 20, height: 20))
-            annotationView?.layer.cornerRadius = annotationView!.bounds.midX
-            annotationView?.layer.borderColor = UIColor.white.cgColor
-            annotationView?.layer.borderWidth = 1
-            annotationView?.layer.shadowColor = UIColor.black.cgColor
-            annotationView?.layer.shadowOpacity = 0.1
-        }
-        
-        annotationView?.backgroundColor = debugAnnotation.color.withAlphaComponent(0.5)
-        
-        return annotationView
+        return debugAnnotationViewFor(debugAnnotation)
     }
     
     func updateShield(for controller: RouteManeuverViewController) {
